backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
redis==5.0.7
rq==1.16.2
httpx==0.27.0
dnspython==2.6.1
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.23.8